feat(可视化): 修改命令名称以及描述信息；
